+++ src/aurelia-authentication.js
@@ -1,14 +1,14 @@
 import {PLATFORM} from 'aurelia-pal';
 import {HttpClient} from 'aurelia-fetch-client';
 import {Config, Rest} from 'aurelia-api';
-import {AuthService} from './authService';
-import {AuthorizeStep} from './authorizeStep';
-import {AuthenticateStep} from './authenticateStep';
 import {BaseConfig} from './baseConfig';
 import {FetchConfig} from './fetchClientConfig';
 import * as LogManager from 'aurelia-logging';
-// import to ensure value-converters get bundled
-import './authFilterValueConverter';
+
+// added for bundling
+import {AuthFilterValueConverter} from './authFilterValueConverter'; // eslint-disable-line no-unused-vars
+import {AuthenticatedFilterValueConverter} from './authenticatedFilterValueConverter'; // eslint-disable-line no-unused-vars
+import {AuthenticatedValueConverter} from './authenticatedValueConverter'; // eslint-disable-line no-unused-vars
 
 /**
  * Configure the plugin.
@@ -16,7 +16,7 @@ import './authFilterValueConverter';
  * @param {{globalResources: Function, container: {Container}}} aurelia
  * @param {{}|Function}                                         config
  */
-function configure(aurelia, config) {
+export function configure(aurelia, config) {
   // ie9 polyfill
   if (!PLATFORM.location.origin) {
     PLATFORM.location.origin = PLATFORM.location.protocol + '//' + PLATFORM.location.hostname + (PLATFORM.location.port ? ':' + PLATFORM.location.port : '');
@@ -68,11 +68,3 @@ function configure(aurelia, config) {
   // Set the client on the config, for use throughout the plugin.
   baseConfig.client = client;
 }
-
-export {
-  configure,
-  FetchConfig,
-  AuthService,
-  AuthorizeStep,
-  AuthenticateStep
-};